Extract methods from view model
